requirements.txt
pydantic
python-dotenv
requests
beautifulsoup4
gspread
google-auth
google-auth-oauthlib
playwright
streamlit
schedule

